requirements/base.txt
# Base requirements

django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
PyJWT==2.8.0
pip-tools==7.3.0
pipdeptree==2.13.0
psycopg2-binary==2.9.7
pillow==10.0.0
channels==4.0.0
daphne==4.0.0